Hope it's this simple
ConfigureStyle is in BuildUI, which is called by __init__ of ReporterGUI, rs_setting_get_font within ConfigureStyle had a `if not self.gui:` clause, when  ConfigureStyle is called self.gui is not set yet as init hasn't finished.
`base.displaygui` determines if ReporterGUI is even called, so it's probably a better choice for this if statement.
Issue #147