Fixes for Issue #150
Found lots of little things here and there that all caused various bad SQL generation, all have been dealt with now.
I'm not sure it's even practical to try and build a regression test for this issue at the moment it would take me weeks and I think it might be better to just get the bug fix release out.